automation project/requirements.txt
selenium==4.15.2
pytest==7.4.3
webdriver-manager==4.0.1
pytest-html==4.1.1
allure-pytest==2.13.2
python-dotenv==1.0.0
